+++ .editorconfig
@@ -1,0 +1,9 @@
+root = true
+
+[*]
+charset                  = utf-8
+end_of_line              = crlf
+indent_style             = space
+indent_size              = 4
+insert_final_newline     = true
+trim_trailing_whitespace = true
