+++ Sidekick.Business/Items/CurrencyItem.cs
@@ -1,0 +1,6 @@
+namespace Sidekick.Business.Items
+{
+    public class CurrencyItem : Item
+    {
+    }
+}
