+++ Sidekick.Business/Items/DivinationCardItem.cs
@@ -1,0 +1,6 @@
+namespace Sidekick.Business.Items
+{
+    public class DivinationCardItem : Item
+    {
+    }
+}
