+++ Sidekick.Business/Items/FragmentItem.cs
@@ -1,0 +1,6 @@
+namespace Sidekick.Business.Items
+{
+    public class FragmentItem : Item
+    {
+    }
+}
