+++ Sidekick.Business/Items/GemItem.cs
@@ -1,0 +1,10 @@
+namespace Sidekick.Business.Items
+{
+    public class GemItem : Item
+    {
+        public string Level { get; set; }
+        public string Quality { get; set; }
+        public bool IsVaalVersion { get; set; }
+        public int ExperiencePercent { get; set; } // percent towards next level
+    }
+}
