+++ Sidekick.Business/Items/Item.cs
@@ -1,0 +1,10 @@
+namespace Sidekick.Business.Items
+{
+    public abstract class Item
+    {
+        public string Name { get; set; }
+        public string Type { get; set; }
+        public string IsCorrupted { get; set; }
+        public string Rarity { get; set; }
+    }
+}
