+++ Sidekick.Business/Items/MapItem.cs
@@ -1,0 +1,11 @@
+namespace Sidekick.Business.Items
+{
+    public class MapItem : Item
+    {
+        public string MapTier { get; set; }
+        public string ItemQuantity { get; set; }
+        public string ItemRarity { get; set; }
+        public string MonsterPackSize { get; set; }
+        public string IsBlight { get; set; }
+    }
+}
