+++ Sidekick.Business/Items/ProphecyItem.cs
@@ -1,0 +1,6 @@
+namespace Sidekick.Business.Items
+{
+    public class ProphecyItem : Item
+    {
+    }
+}
