+++ Sidekick.Business/Localization/LanguageEnum.cs
@@ -1,0 +1,14 @@
+namespace Sidekick.Business.Localization
+{
+    public enum LanguageEnum
+    {
+        English,
+        French,
+        German,
+        Korean,
+        Portuguese,
+        Russian,
+        Spanish,
+        Thai
+    };
+}
