+++ Sidekick.Business/Loggers/Log.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Sidekick.Business.Loggers
+{
+    public class Log
+    {
+        public DateTime Date { get; set; }
+        public string Message { get; set; }
+        public LogState State { get; set; }
+    }
+}
