+++ Sidekick.Business/Loggers/LogState.cs
@@ -1,0 +1,10 @@
+namespace Sidekick.Business.Loggers
+{
+    public enum LogState
+    {
+        None,
+        Success,
+        Error,
+        Warning
+    }
+}
