+++ Sidekick.Core/DependencyInjection/Services/LifetimeEnum.cs
@@ -1,0 +1,9 @@
+namespace Sidekick.Core.DependencyInjection.Services
+{
+  public enum LifetimeEnum
+  {
+    Scoped,
+    Singleton,
+    Transcient
+  }
+}
