+++ Sidekick.Core/DependencyInjection/Startup/ISidekickStartup.cs
@@ -1,0 +1,10 @@
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Sidekick.Core.DependencyInjection
+{
+  public interface ISidekickStartup
+  {
+    void ConfigureServices(IServiceCollection services, IConfiguration configuration);
+  }
+}
