+++ Sidekick.Input/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Sidekick.Input
+{
+  public class Class1
+  {
+  }
+}
